soclinq_backend/backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
python-dotenv
gunicorn
djangorestframework-simplejwt
requests
channels
channels-redis
django-ratelimit


